# Conflicts:
#	jims-service/src/main/java/com/jims/lab/service/LabTestMasterServiceImpl.java

+++ jims-api/src/main/java/com/jims/exam/entity/Orders.java
@@ -25,8 +25,8 @@ public class Orders extends DataEntity<Orders> {
 	private String visitId;		// 病人本次住院标识
 	private String orgId;    //组织机构id
 	private String clinicId;		//就诊id
-	private Long orderNo;		// 医嘱序号
-	private Long orderSubNo;		// 医嘱子序号
+	private int orderNo;		// 医嘱序号
+	private int orderSubNo;		// 医嘱子序号
 	private String repeatIndicator;		// 长期医嘱标志
 	private String orderClass;		// 医嘱类别
 	private String orderText;		// 医嘱正文
@@ -127,22 +127,22 @@ public class Orders extends DataEntity<Orders> {
 		this.visitId = visitId;
 	}
 
-	public Long getOrderNo() {
+	public int getOrderNo() {
 		return orderNo;
 	}
 
-	public void setOrderNo(Long orderNo) {
+	public void setOrderNo(int orderNo) {
 		this.orderNo = orderNo;
 	}
-	
-	public Long getOrderSubNo() {
+
+	public int getOrderSubNo() {
 		return orderSubNo;
 	}
 
-	public void setOrderSubNo(Long orderSubNo) {
+	public void setOrderSubNo(int orderSubNo) {
 		this.orderSubNo = orderSubNo;
 	}
-	
+
 	@Length(min=0, max=1, message="长期医嘱标志长度必须介于 0 和 1 之间")
 	public String getRepeatIndicator() {
 		return repeatIndicator;
